Rename generators to match generator IDs.
