+++ pipeline/matching/engines/config.py
@@ -1,4 +1,4 @@
 from pipeline.matching.engines.main import MainMatchingEngine
-from pipeline.matching.engines.silly import SillyMatchingEngine
+from pipeline.matching.engines.simple import SimpleMatchingEngine
 
-ENGINES = {"main": MainMatchingEngine, "silly": SillyMatchingEngine}
+ENGINES = {"main": MainMatchingEngine, "simple": SimpleMatchingEngine}
